controllers fixing, users remained

+++ controllers/productType.controller.js
@@ -5,9 +5,15 @@ const Op = db.Sequelize.Op;
 // Create and Save a new ProductType
 exports.create = (req, res) => {
   // Validate request
-  if (!req.body.title) {
+  if (!req.body.name) {
     res.status(400).send({
-      message: "Content can not be empty!"
+      message: "name can not be empty!"
+    });
+    return;
+  }
+  if (!req.body.createdBy) {
+    res.status(400).send({
+      message: "createdBy can not be empty!"
     });
     return;
   }
@@ -15,7 +21,7 @@ exports.create = (req, res) => {
   // Create a ProductType
   const productType = {
     name: req.body.name,
-    createdBy: req.body.createdBy ?? null,
+    createdBy: req.body.createdBy,
     updatedBy: null
   };
 
@@ -67,6 +73,12 @@ exports.findOne = (req, res) => {
 
 // Update a ProductType by the id in the request
 exports.update = (req, res) => {
+  if (!req.body.updatedBy) {
+    res.status(400).send({
+      message: "updatedBy can not be empty!"
+    });
+    return;
+  }
   const id = req.params.id;
 
   ProductType.update(req.body, {
@@ -92,6 +104,12 @@ exports.update = (req, res) => {
 
 // Delete a ProductType with the specified id in the request
 exports.delete = (req, res) => {
+  if (!req.body.updatedBy) {
+    res.status(400).send({
+      message: "updatedBy can not be empty!"
+    });
+    return;
+  }
   const id = req.params.id;
 
   ProductType.destroy({
